MEDCAL/requirements.txt
matplotlib==3.10.8
numpy==2.4.1
pandas==2.3.3
Pillow==12.1.0
plotly==6.5.2
scikit_learn==1.8.0
seaborn==0.13.2
streamlit==1.51.0
